+++ src/ui/pages/product/SingleProduct.jsx
@@ -1,0 +1,10 @@
+import React from 'react'
+import './SingleProduct.scss'
+
+const SingleProduct = () => {
+  return (
+    <div>Single Product page</div>
+  )
+}
+
+export default SingleProduct
