+++ src/ui/pages/product/SingleProduct.scss
@@ -1,0 +1,11 @@
+@import "../../../scss/base";
+
+.product_wrapper {
+    padding: 0 0;
+}
+
+.product_section {
+    margin: 0 auto;
+    max-width: $page-width-xl;
+    padding: 2rem 3rem;
+}
